src/requirements.txt
pydub>=0.25.1
moviepy>=1.0.3
numpy>=1.21.0
scipy>=1.7.0
pycryptodome>=3.15.0
sounddevice>=0.4.4
fastapi>=0.68.0
uvicorn>=0.15.0
pytest>=6.2.4
pytest-mock>=3.6.1
pytest-asyncio>=0.15.1
ffmpeg-python>=0.2.0
python-multipart>=0.0.5
fastmcp>=2.0.0
click>=8.0.0
typer>=0.9.0
rich>=13.0.0